+++ src/app/containers/user/user.component.ts
@@ -5,7 +5,6 @@ import { EchoesState } from '@core/store';
 import * as UserProfile from '@core/store/user-profile/user-profile.selectors';
 import { AppApi } from '@api/app.api';
 
-
 @Component({
   selector: 'app-user',
   encapsulation: ViewEncapsulation.None,
@@ -32,14 +31,9 @@ export class UserComponent implements OnInit {
   currentPlaylist$ = this.store.select(UserProfile.getUserViewPlaylist);
   isSignedIn$ = this.store.select(UserProfile.getIsUserSignedIn);
 
-  constructor(
-    private appApi: AppApi,
-    public store: Store<EchoesState>
-  ) { 
-    console.log('LAZY..');
-  }
+  constructor(private appApi: AppApi, public store: Store<EchoesState>) {}
 
-  ngOnInit() { }
+  ngOnInit() {}
 
   signInUser() {
     this.appApi.signinUser();