+++ src/app/core/store/now-playlist/now-playlist.actions.ts
@@ -22,92 +22,94 @@ export enum ActionTypes {
   LOAD_PLAYLIST_END = '[NowPlaylist] LOAD_PLAYLIST_END',
   PLAY_PLAYLIST = '[NowPlaylist] PLAY_PLAYLIST',
   PLAY_PLAYLIST_START = '[NowPlaylist] PLAY_PLAYLIST_START',
-  PLAYER_STATE_CHANGE = '[NowPlaylist] PLAYER_STATE_CHANGE',
+  PLAYER_STATE_CHANGE = '[NowPlaylist] PLAYER_STATE_CHANGE'
 }
 
 export class SeekTo implements Action {
   public type = ActionTypes.SELECT_AND_SEEK_TO_TIME;
-  constructor(public payload: { time: string; media: GoogleApiYouTubeVideoResource }){}
+  constructor(
+    public payload: { time: string; media: GoogleApiYouTubeVideoResource }
+  ) {}
 }
 export class QueueLoadVideo implements Action {
   public type = ActionTypes.QUEUE_LOAD_VIDEO;
-  constructor(public payload: GoogleApiYouTubeVideoResource) { }
+  constructor(public payload: GoogleApiYouTubeVideoResource) {}
 }
 
 export class UpdateIndexByMedia implements Action {
   public type = ActionTypes.UPDATE_INDEX;
-  constructor(public payload: string) { }
+  constructor(public payload: string) {}
 }
 
 export class QueueFailed implements Action {
   public type = ActionTypes.QUEUE_FAILED;
-  constructor(public payload: GoogleApiYouTubeVideoResource) { }
+  constructor(public payload: GoogleApiYouTubeVideoResource) {}
 }
 export class QueueVideo implements Action {
   public type = ActionTypes.QUEUE;
-  constructor(public payload: GoogleApiYouTubeVideoResource) { }
+  constructor(public payload: GoogleApiYouTubeVideoResource) {}
 }
 export class QueueVideos implements Action {
   public type = ActionTypes.QUEUE_VIDEOS;
-  constructor(public payload: GoogleApiYouTubeVideoResource[]) { }
+  constructor(public payload: GoogleApiYouTubeVideoResource[]) {}
 }
 export class RemoveVideo implements Action {
   public type = ActionTypes.REMOVE;
-  constructor(public payload: GoogleApiYouTubeVideoResource) { }
+  constructor(public payload: GoogleApiYouTubeVideoResource) {}
 }
 
 export class FilterChange implements Action {
   public type = ActionTypes.FILTER_CHANGE;
-  constructor(public payload: string) { }
+  constructor(public payload: string) {}
 }
 export class SelectVideo implements Action {
   public type = ActionTypes.SELECT;
-  constructor(public payload: GoogleApiYouTubeVideoResource) { }
+  constructor(public payload: GoogleApiYouTubeVideoResource) {}
 }
 
 export class PlayPlaylistAction implements Action {
   readonly type = ActionTypes.PLAY_PLAYLIST;
-  constructor(public payload: string) { }
+  constructor(public payload: string) {}
 }
 export class PlayPlaylistStartAction implements Action {
   readonly type = ActionTypes.PLAY_PLAYLIST_START;
-  constructor(public payload: GoogleApiYouTubeVideoResource) { }
+  constructor(public payload: GoogleApiYouTubeVideoResource) {}
 }
 
 export class LoadPlaylistAction implements Action {
   readonly type = ActionTypes.LOAD_PLAYLIST_START;
-  constructor(public payload: string) { }
+  constructor(public payload: string) {}
 }
 
 export class LoadPlaylistEndAction implements Action {
   readonly type = ActionTypes.LOAD_PLAYLIST_END;
-  constructor(public payload: GoogleApiYouTubeVideoResource[]) { }
+  constructor(public payload: GoogleApiYouTubeVideoResource[]) {}
 }
 
 export class MediaEnded implements Action {
   public type = ActionTypes.MEDIA_ENDED;
-  constructor(public payload?: any) { }
+  constructor(public payload?: any) {}
 }
 
 export class SelectNext implements Action {
   public type = ActionTypes.SELECT_NEXT;
-  constructor(public payload?: any) { }
+  constructor(public payload?: any) {}
 }
 export class SelectPrevious implements Action {
   public type = ActionTypes.SELECT_PREVIOUS;
-  constructor(public payload?: any) { }
+  constructor(public payload?: any) {}
 }
 export class RemoveAll implements Action {
   public type = ActionTypes.REMOVE_ALL;
-  constructor(public payload?: any) { }
+  constructor(public payload?: any) {}
 }
 export class ToggleRepeat implements Action {
   public type = ActionTypes.TOGGLE_REPEAT;
-  constructor(public payload = '') { }
+  constructor(public payload = '') {}
 }
 export class PlayerStateChange implements Action {
   public type = ActionTypes.PLAYER_STATE_CHANGE;
-  constructor(public payload: YT.OnStateChangeEvent) { }
+  constructor(public payload: YT.PlayerState) {}
 }
 
 export type Actions =